+++ src/templates.rs
@@ -31,7 +31,7 @@ mod output;
 mod preview;
 mod range;
 mod rare;
-mod sat;
+pub mod sat;
 mod transaction;
 
 #[derive(Boilerplate)]
